refactor: unify KPI card styling across modules
Extract reusable KPICard component to modulos/kpi_card.py.
Both clientes and inventario modules now use identical theme-aware
card with accent bar, typography tokens, and dynamic color variants.

Changes:
- Create modulos/kpi_card.py: shared KPICard class with design tokens
  (no hardcoded colors, respects theme via Colors token)
- Update modulos/clientes.py: replace hardcoded stats bar with KPICard grid
- Update modulos/inventario_local.py: use shared KPICard, alias _InvKPICard
- Both modules now have consistent KPI styling, spacing, and theming

Benefits:
- Single source of truth for KPI styling
- Automatic theme support (dark/light) via design tokens
- Consistent spacing, typography, and color palette
- Easy to extend with new modules
